# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Music Service",
...
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "YouTube Music API Service",
      "url": "https://github.com/your-repo"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/explore/": {
      "get": {
        "tags": [
          "explore",
          "explore"
        ],
        "summary": "Get explore content",
        "description": "Obtiene contenido de exploración: moods, géneros, charts y home. Incluye top songs y trending con stream URLs opcionales.",
        "operationId": "explore_music_api_v1_explore__get",
        "parameters": [
          {
            "name": "include_stream_urls",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Incluir stream URLs y mejores thumbnails para charts",
              "default": false,
              "title": "Include Stream Urls"
            },
            "description": "Incluir stream URLs y mejores thumbnails para charts"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Número máximo de canciones en charts",
              "default": 10,
              "title": "Limit"
            },
            "description": "Número máximo de canciones en charts"
          },
          {
            "name": "start_index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Índice inicial para paginación de charts",
              "default": 0,
              "title": "Start Index"
            },
            "description": "Índice inicial para paginación de charts"
          },
          {
            "name": "prefetch_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": -1,
              "description": "Número de URLs a obtener en paralelo (0=none, -1=todas)",
              "default": 10,
              "title": "Prefetch Count"
            },
            "description": "Número de URLs a obtener en paralelo (0=none, -1=todas)"
          }
        ],
        "responses": {
          "200": {
            "description": "Contenido de exploración obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Explore Music Api V1 Explore  Get"
                },
                "example": {
                  "moods_genres": [
                    {
                      "title": "Cumbia",
                      "params": "ggMPOg1uX3hRRFdlaEhHU09k"
                    }
                  ],
                  "home": [],
                  "charts": {
                    "top_songs": [
                      {
                        "videoId": "rMbATaj7Il8",
                        "title": "Song Title",
                        "artists": [
                          {
                            "name": "Artist"
                          }
                        ],
                        "stream_url": "https://...",
                        "thumbnail": "https://..."
                      }
                    ],
                    "trending": []
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/moods": {
      "get": {
        "tags": [
          "explore",
          "explore"
        ],
        "summary": "Get mood categories",
        "description": "Obtiene todas las categorías de moods y géneros disponibles en YouTube Music.",
        "operationId": "get_mood_categories_api_v1_explore_moods_get",
        "responses": {
          "200": {
            "description": "Categorías obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Mood Categories Api V1 Explore Moods Get"
                },
                "example": {
                  "categories": {
                    "For you": [
                      {
                        "params": "ggMPOg1uX1ZwN0pHT2NBT1Fk",
                        "title": "1980s"
                      }
                    ],
                    "Genres": [
                      {
                        "params": "ggMPOg1uX3hRRFdlaEhHU09k",
                        "title": "Cumbia"
                      }
                    ],
                    "Moods & moments": [
                      {
                        "params": "ggMPOg1uXzVuc0dnZlhpV3Ba",
                        "title": "Chill"
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/v1/explore/moods/{params}": {
      "get": {
        "tags": [
          "explore",
          "explore"
        ],
        "summary": "Get mood/genre playlists",
        "description": "Obtiene playlists de una categoría de mood o género usando sus parámetros codificados.",
        "operationId": "get_mood_playlists_api_v1_explore_moods__params__get",
        "parameters": [
          {
            "name": "params",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Parámetros codificados de la categoría",
              "examples": {
                "example1": {
                  "value": "ggMPOg1uX3hRRFdlaEhHU09k"
                }
              },
              "title": "Params"
            },
            "description": "Parámetros codificados de la categoría"
          }
        ],
        "responses": {
          "200": {
            "description": "Playlists obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Mood Playlists Api V1 Explore Moods  Params  Get"
                },
                "example": {
                  "playlists": [
                    {
                      "title": "Cumbia Mix",
                      "playlistId": "PL123"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Categoría no encontrada (params inválidos)"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/explore/charts": {
      "get": {
        "tags": [
          "explore",
          "explore"
        ],
        "summary": "Get music charts",
        "description": "Obtiene los charts de YouTube Music: top songs y trending. Opcionalmente incluye stream URLs.",
        "operationId": "get_charts_api_v1_explore_charts_get",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Código de país ISO 3166-1 Alpha-2 (ej: 'US', 'PE'). Default: global",
              "title": "Country"
            },
            "description": "Código de país ISO 3166-1 Alpha-2 (ej: 'US', 'PE'). Default: global"
          },
          {
            "name": "include_stream_urls",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Incluir stream URLs y mejores thumbnails",
              "default": false,
              "title": "Include Stream Urls"
            },
            "description": "Incluir stream URLs y mejores thumbnails"
          }
        ],
        "responses": {
          "200": {
            "description": "Charts obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Charts Api V1 Explore Charts Get"
                },
                "example": {
                  "top_songs": [
                    {
                      "videoId": "rMbATaj7Il8",
                      "title": "Top Song",
                      "artists": [
                        {
                          "name": "Artist"
                        }
                      ],
                      "stream_url": "https://...",
                      "thumbnail": "https://..."
                    }
                  ],
                  "trending": [],
                  "country": "global"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}